refactor: moved theme extension generators to their own folder
