+++ lib/src/data/generators/theme_extension_generators/number_theme_extension_generator.dart
@@ -1,5 +1,5 @@
 import 'package:code_builder/code_builder.dart';
-import 'package:figmage/src/data/generators/values_by_mode_theme_extension_generator.dart';
+import 'package:figmage/src/data/generators/theme_extension_generators/values_by_mode_theme_extension_generator.dart';
 
 /// {@template number_theme_extension_generator}
 /// A [ValuesByModeThemeExtensionGenerator] for numbers from doubles.
