+++ test/full_coverage_test.dart
@@ -8,14 +8,14 @@ import 'package:figmage/src/commands/shared/arg_names.dart';
 import 'package:figmage/src/commands/shared/forge_settings_providers.dart';
 import 'package:figmage/src/commands/shared/generation_notifier.dart';
 import 'package:figmage/src/commands/update/update_command.dart';
-import 'package:figmage/src/data/generators/color_theme_extension_generator.dart';
+import 'package:figmage/src/data/generators/theme_extension_generators/color_theme_extension_generator.dart';
 import 'package:figmage/src/data/generators/generator_util.dart';
-import 'package:figmage/src/data/generators/number_theme_extension_generator.dart';
+import 'package:figmage/src/data/generators/theme_extension_generators/number_theme_extension_generator.dart';
 import 'package:figmage/src/data/generators/padding_generator.dart';
 import 'package:figmage/src/data/generators/spacer_generator.dart';
-import 'package:figmage/src/data/generators/text_style_theme_extension_generator.dart';
+import 'package:figmage/src/data/generators/theme_extension_generators/text_style_theme_extension_generator.dart';
 import 'package:figmage/src/data/generators/value_names_theme_class_generator.dart';
-import 'package:figmage/src/data/generators/values_by_mode_theme_extension_generator.dart';
+import 'package:figmage/src/data/generators/theme_extension_generators/values_by_mode_theme_extension_generator.dart';
 import 'package:figmage/src/data/repositories/dart_code_file_writer_repository.dart';
 import 'package:figmage/src/data/repositories/dart_post_generation_repository.dart';
 import 'package:figmage/src/data/repositories/figma_styles_repository.dart';
